api/requirements.txt
fastapi
uvicorn
motor
passlib
bcrypt
python-multipart
pyjwt
starlette
requests
python-jose[cryptography]
nh3
cryptography
rsa